Refactored to split AWS, DB, and validation related functions into their own files.
