+++ server/aws.js
@@ -1,0 +1,55 @@
+/*
+
+  ___ ___       ___               _ _       _   ___ ___ 
+ | __/ __|  ___| _ \___ _ _ _ __ (_) |_    /_\ | _ \_ _|
+ | _|\__ \ / -_)  _/ -_) '_| '  \| |  _|  / _ \|  _/| | 
+ |_| |___/ \___|_| \___|_| |_|_|_|_|\__| /_/ \_\_| |___|
+
+*/
+
+//*******************************************************************
+
+'use strict';
+
+//*******************************************************************
+// required modules
+
+const AWS = require('aws-sdk');
+
+//*************************************************************
+// AWS
+
+const AWS_ACCESS_KEY_ID = process.env.AWS_ACCESS_KEY_ID;
+const AWS_SECRET_ACCESS_KEY = process.env.AWS_SECRET_ACCESS_KEY;
+const AWS_REGION = process.env.AWS_REGION;
+const AWS_BUCKET_NAME = process.env.AWS_BUCKET_NAME;
+
+AWS.config.update({
+	accessKeyId: AWS_ACCESS_KEY_ID,
+	secretAccessKey: AWS_SECRET_ACCESS_KEY,
+	region: AWS_REGION
+});
+
+const s3 = new AWS.S3();
+
+//*************************************************************
+
+function uploadFile(fileInfo, callback){
+	const params = {
+		Bucket: AWS_BUCKET_NAME, 
+		Key: fileInfo.keyname,
+		Body: fileInfo.buffer,
+		ACL: 'private' 
+	};
+
+	s3.putObject(params, function(err, data) {
+		if (err) {
+			return callback(err, null);
+		}
+		else {     
+			return callback(null, data);
+		}      
+	});
+}
+
+module.exports.uploadFile = uploadFile;
